Enhancement: make all hex-bound overloaded operators elemental
I wanted also to move their declaration inside the type definition, so
to avoid exporting all the operator interfaces along with the type, with
a serious risk of forgetting about them if one includes the module with
the < only: hex > attribute. But I ran into nasty problems with the add,
sub, dot, lhs, rhs functions: being proper object-oriented methods does
require the intent(in) dummy variables to have class(hex) declaration,
but the result cannot have the same, being not a dummy variable. Whilst
a subroutine would not be feasible to be overloaded to such an operator.
So in the end we either accept that a class(hex) --> type(hex) function
would perform a type conversion (mmmh) or give up on the OO way to bind
the methods to the type. So the real question is: how we can get a full
OO-style polymorphic binary operators?

> A TYPEOF() intrinsic would come so handy for the result declaration,
  right? There is actually a proposal for the upcoming F2023 standard,
  but I'm not sure it would really address this issue.
  Ref to: https://github.com/j3-fortran/fortran_proposals/issues/158