Extend `honeyplots` module to handle `xy_lattice` types
Now there is a generic interface (plot) overloaded to both hex_plot and
the new xy_plot. The latter takes all the same optional backend-related
input variables, but a type(xy_lattice) instead of an array of type(hex)

Furthermore it takes optionally NN_mask and NNN_mask, as generated using
the xy_neighbors module. If the former is passed it draws the NN bonds,
if the latter is passed it draws dashed NNN links. If nothing is passed
it draws only the lattice points. PLEASE NOTE THAT THE BOND/LINK OPTION
IS NOT AVAILABLE WITH THE GNUPLOT BACKEND. I don't know how to get this
working with the gnuplot interface provided by OGPF, since it allows to
generate all plot-parts in the same call, which is not feasible if you
want to plot single segments, I think. Maybe we should just drop OGPF.

> This allowed to test both the NN and the NNN masks, for the very basic
  structure we currently generate in the test script. It works perfectly.

TODO: add finally a module to automatically generate the most commonly
      needed shapes: generic radius flakes, rectangular strips, common
      parallelogram-shaped supercells (what PythTB does...), triangles.