Remove `xy_tile` type and related `corner2lattice` type conversion
We can store tiles in xy_lattice type instances, so there is really no
need for a dedicated type for static collection of N=6 lattice sites.

Additionally we add a get_sublattice(lattice,label) public function to
xy_coordinates module, to help the user avoid to take dangerous guesses
about their indexing. For the simple example we have apparently "A" and
"B" sites are effectively on odd and even indices, but I suspect this
could be not true for more general structures. Incidentally this solves
some gfortran warnings about temporary array allocations, where we were
doing:

         ! Sublattice "A"
         MaskA = all_sites%label=="A"
         Ax = pack(all_sites%x,MaskA)
         Ay = pack(all_sites%y,MaskA)

         ! Sublattice "B"
         MaskB = all_sites%label=="B"
         Bx = pack(all_sites%x,MaskB)
         By = pack(all_sites%y,MaskB)

Now we do instead:

         sublattice = get_sublattice(lattice,"A")
         Ax = sublattice%site%x
         Ay = sublattice%site%y

         sublattice = get_sublattice(lattice,"B")
         Bx = sublattice%site%x
         By = sublattice%site%y

Overall I think the new version is both more readable (and this is
paramount, where the user has to extract the sublattice himself!).

E.g. a user can fill a staggered onsite term in the Hamiltonian as:

         sublattice = get_sublattice(lattice,"A")
         indices = sublattice%site%key
         do i = 1,size(indices)
            H(indices(i),indices(i)) = + m
         enddo

         sublattice = get_sublattice(lattice,"B")
         indices = sublattice%site%key
         do i = 1,size(indices)
            H(indices(i),indices(i)) = - m
         enddo

which has been added to the test script.