Remove dependency on `stdlib_hash_32bit` + CI :: remove Qt installation
I don't wand to spend time now to understand why it's making the CI fail.
Also maybe it's too much to just build a quick test-script / example.

I roll instead my own random integer generator, as a local function
to the test program. It's simple, but it works.

We also prune the Qt installation step in CI: we can't really use the Qt
terminal in gnuplot and it takes so much time in the CI run.