Add `xy_nn_hop` and `xy_nnn_hop` within `xy_neighbors` module
These public pure functions take a xy_site and return one of its NNs or
NNNs (respectively), provided an integer i ∈ ℤ and a unit-cell object.

For NNs there are only 3 inequivalent results, depending on the label of
the site: "A" activates some directions, "B" others, they correspond to
respectively i=1,3,5 and i=2,4,6, with reference to ith_corner_offset.

For NNNs there are instead 6 inequivalent results, for whatever site.

This makes the library much, much, much more flexible whenever not only
the shells are needed, but also the "direction" of a link is relevant.

> E.g. in a Haldane model you need to know whether you are hopping in
  clock-wise or anti-clock-wise direction to assign the correct phase
  in the hamiltonian. With only the NNN-mask you cannot determine how
  to assign the phase for that each NNN link is stored symmetrically.
  If you instead loop over the xy_lattice and, for each xy_site, find
  all the NNNs with xy_nnn_hop(layout,site,i) you can correctly match
  the phase by just partitioning the {i} values: odd would correspond
  to a phase and even to the opposite, depending on the site label.

A suitable unit-test script has been added to the test-suite.