API update: promote `xy_site` and `hex_orientation` to top-level module
Recent work with the Kane-Mele model has made clear that we need quite
often to manually generate the hoppings, as well as defining a custom
unit-cell, already at the unit-vector level. It follow that both types
xy_site and hex_orientation are needed for a fairly basic usage.

Since we change API (even if in a non-breaking fashion) we should soon
bump the version. Though I'll wait until I update the README.