Failed try at custom plot colors
We wanted to leave behind the default matplotlib colors, rather plotting
with the two colors used in the logo (official fortran-lang color + our
custom yellow).

While this is trivial with OGPF with pyplot-fortran we find some tricky
issues, which are documented at:

     > https://github.com/jacobwilliams/pyplot-fortran/issues/37

Unfortunately this means the update would remain frozen here in the fork
until we find a solution.