FIX fpm build
From fpm release v0.9.0 the stdlib becomes an officially supported
"metapackage". As such it cannot be included with a github link and
a branch, as before (and as currently documented in their readme…).
The new, simpler, toml syntax is here implemented, so to fix the
build of HoneyTools on the github runner. Consequently we add a min
version to fpm in the README (as annoying as it looks, recall that
fpm is still in alpha stage, so this would happen a lot).